Project restructure
- Separate modules into folders and namespaces
