+++ src/Nethermind/Nethermind.Tools.Kute/Auth/IAuth.cs
@@ -1,9 +1,9 @@
 // SPDX-FileCopyrightText: 2023 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
-namespace Nethermind.Tools.Kute;
+namespace Nethermind.Tools.Kute.Auth;
 
-interface ISystemClock
+interface IAuth
 {
-    DateTimeOffset UtcNow { get; }
+    string AuthToken { get; }
 }
