+++ src/Nethermind/Nethermind.Tools.Kute/SystemClock/ISystemClock.cs
@@ -1,9 +1,9 @@
 // SPDX-FileCopyrightText: 2023 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
-namespace Nethermind.Tools.Kute;
+namespace Nethermind.Tools.Kute.SystemClock;
 
-interface IAuth
+interface ISystemClock
 {
-    string AuthToken { get; }
+    DateTimeOffset UtcNow { get; }
 }
