+++ src/Nethermind/Nethermind.Tools.Kute/SystemClock/SystemClock.cs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
-namespace Nethermind.Tools.Kute;
+namespace Nethermind.Tools.Kute.SystemClock;
 
 class SystemClock : ISystemClock
 {
